TRIPLEX-838 List stories
