+++ stories/List/ListItemControls.stories.tsx
@@ -1,15 +1,14 @@
 import React from "react";
-import { StoryObj } from "@storybook/react";
-import { Controls, Description, Stories, Subtitle, Title } from "@storybook/addon-docs/blocks";
-import { ListItemControls, ListItemControlsButtonDropdown, ListItemControlsButton } from "../../src/components/List";
-import { AttachmentStrokeSrvIcon20, DotshorizontalStrokeSrvIcon20 } from "@sberbusiness/icons-next";
+import { Meta, StoryObj } from "@storybook/react";
+import { Description, Stories, Title } from "@storybook/addon-docs/blocks";
+import { ListItemControls } from "@sberbusiness/triplex-next";
+import { DefaultExample, DefaultExampleSource } from "./examples/ListItemControls";
 
-export default {
+const meta = {
     title: "Components/List/ListItemControls",
     component: ListItemControls,
     tags: ["autodocs"],
     parameters: {
-        testRunner: { skip: true },
         docs: {
             description: {
                 component: `
@@ -19,83 +18,25 @@ export default {
             page: () => (
                 <>
                     <Title />
-                    <Subtitle />
                     <Description />
-                    <Controls of={Default} />
                     <Stories />
                 </>
             ),
         },
     },
-} as const;
+} satisfies Meta<typeof ListItemControls>;
+
+export default meta;
 
 export const Default: StoryObj<typeof ListItemControls> = {
-    render: () => {
-        const options = [
-            {
-                id: "list-controls-dropdown-option-1",
-                label: "Текст пункта меню 1",
-                onSelect: () => alert("Выбран пункт меню 1."),
-            },
-            {
-                id: "list-controls-dropdown-option-2",
-                label: "Текст пункта меню 2",
-                onSelect: () => alert("Выбран пункт меню 2."),
-            },
-            {
-                id: "list-controls-dropdown-option-3",
-                label: "Текст пункта меню 3",
-                onSelect: () => alert("Выбран пункт меню 3."),
-            },
-        ];
-        return (
-            <>
-                <ListItemControls>
-                    <ListItemControlsButton icon={<AttachmentStrokeSrvIcon20 paletteIndex={0} />}>
-                        Скачать
-                    </ListItemControlsButton>
-                    <ListItemControlsButtonDropdown
-                        icon={<DotshorizontalStrokeSrvIcon20 paletteIndex={0} />}
-                        options={options}
-                    >
-                        Еще
-                    </ListItemControlsButtonDropdown>
-                </ListItemControls>
-
-                <br />
-
-                <ListItemControls style={{ height: "100px" }}>
-                    <ListItemControlsButton icon={<AttachmentStrokeSrvIcon20 paletteIndex={0} />}>
-                        Скачать
-                    </ListItemControlsButton>
-                    <ListItemControlsButtonDropdown
-                        icon={<DotshorizontalStrokeSrvIcon20 paletteIndex={0} />}
-                        options={options}
-                    >
-                        Длинная подпись
-                    </ListItemControlsButtonDropdown>
-                </ListItemControls>
-
-                <br />
-
-                <ListItemControls>
-                    <ListItemControlsButton icon={<AttachmentStrokeSrvIcon20 paletteIndex={0} />} />
-                    <ListItemControlsButtonDropdown
-                        icon={<DotshorizontalStrokeSrvIcon20 paletteIndex={0} />}
-                        options={options}
-                    />
-                </ListItemControls>
-
-                <br />
-
-                <ListItemControls>
-                    <ListItemControlsButton>Кнопка</ListItemControlsButton>
-                    <ListItemControlsButtonDropdown options={options}>Дропдаун</ListItemControlsButtonDropdown>
-                </ListItemControls>
-            </>
-        );
-    },
+    render: DefaultExample,
     parameters: {
         controls: { disable: true },
+        docs: {
+            source: {
+                code: DefaultExampleSource,
+                language: "tsx",
+            },
+        },
     },
 };
