+++ stories/List/examples/List/DefaultExample.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { List, ListItem } from "@sberbusiness/triplex-next";
+
+export const DefaultExample = () => (
+    <div style={{ width: "500px" }}>
+        <List>
+            <ListItem>Элемент списка</ListItem>
+            <ListItem>Элемент списка</ListItem>
+            <ListItem>Элемент списка</ListItem>
+        </List>
+    </div>
+);
