+++ stories/List/examples/ListItem/LoadingExample.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { List, ListItem, ListItemLoading } from "@sberbusiness/triplex-next";
+
+export const LoadingExample = () => (
+    <div style={{ maxWidth: "500px" }}>
+        <List>
+            <ListItem>
+                <ListItemLoading />
+            </ListItem>
+        </List>
+    </div>
+);
