+++ stories/List/examples/ListItem/SelectableExample.tsx
@@ -1,0 +1,18 @@
+import React, { useState } from "react";
+import { List, ListItem, ListItemContent, ListItemSelectable } from "@sberbusiness/triplex-next";
+
+export const SelectableExample = () => {
+    const [selected, setSelected] = useState(false);
+
+    return (
+        <div style={{ maxWidth: "500px" }}>
+            <List>
+                <ListItem>
+                    <ListItemSelectable selected={selected} onSelect={setSelected}>
+                        <ListItemContent>Элемент списка</ListItemContent>
+                    </ListItemSelectable>
+                </ListItem>
+            </List>
+        </div>
+    );
+};
